toString test updated with fixtures
