+++ Tests/Fixtures/Entity/FooGetName.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Sonata\DashboardBundle\Tests\Fixtures\Entity;
+
+class FooGetName
+{
+    public function getName()
+    {
+        return 'GetName';
+    }
+}
