+++ Tests/Fixtures/Entity/FooGetNameNull.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Sonata\DashboardBundle\Tests\Fixtures\Entity;
+
+class FooGetNameNull
+{
+    public function getName()
+    {
+        return;
+    }
+
+    public function __toString()
+    {
+        return 'GetNameNull';
+    }
+}
