+++ Tests/Fixtures/Entity/FooNoGetName.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Sonata\DashboardBundle\Tests\Fixtures\Entity;
+
+class FooNoGetName
+{
+    public function __toString()
+    {
+        return 'NoGetName';
+    }
+}
